+++ src/components/SearchForm.jsx
@@ -1,16 +1,7 @@
-import { useCallback, useEffect, useState } from 'react'
-import { useAtom } from 'jotai'
-
 import styled from 'styled-components'
 import { BsSearch, BsStar } from 'react-icons/bs'
 
-import { useSearchParams } from 'react-router-dom'
-
-import { fetchSearchByInput } from '../Utils/fetchers/fetchSearchByInput'
-import getRandomCharacter from '../Utils/getRandomCharacter'
-
-import { charactersResults, matchingResults } from '../atoms'
-import { fetchRandomCharacter } from '../Utils/fetchers/fetchRandomCharacter'
+import useFetchCharacters from '../hooks/useFetchCharacters'
 
 const StyledForm = styled.form`
   width: 70%;
@@ -54,87 +45,8 @@ const FavoriteStar = styled(BsStar)`
   filter: opacity(15%);
 `
 
-const api = 'http://gateway.marvel.com/v1/public/characters?'
-const apiKey = 'f4e63a51401e5c498e1740d446ae8f5d'
-
 export const SearchForm = () => {
-  const [inputString, setInputString] = useState('')
-  const [isSubmitted, setIsSubmitted] = useState(false)
-  const [, setSearchParams] = useSearchParams('')
-
-  const [, setCardsData] = useAtom(charactersResults)
-  const [, setResultsList] = useAtom(matchingResults)
-
-  const handleFetchRandom = useCallback(async () => {
-    const query = getRandomCharacter()
-    const results = await fetchRandomCharacter({
-      api,
-      apiKey,
-      query,
-      limit: 9
-    })
-
-    setCardsData(results)
-    setSearchParams({ character: query })
-  })
-
-  useEffect(() => {
-    handleFetchRandom()
-  }, [])
-
-  const handleFetchByInput = useCallback(async (query) => {
-    if (query !== '') {
-      const results = await fetchSearchByInput({
-        api,
-        apiKey,
-        query,
-        limit: 9
-      })
-
-      setResultsList(results)
-      setCardsData(results)
-      setSearchParams({ character: inputString })
-    }
-  }, [])
-
-  useEffect(() => {
-    handleFetchByInput(inputString)
-  }, [isSubmitted])
-
-  const handleFetchMatchingResults = useCallback(async (query) => {
-    if (query !== '') {
-      const results = await fetchSearchByInput({
-        api,
-        apiKey,
-        query,
-        limit: 9
-      })
-      setResultsList(results)
-    }
-  }, [])
-
-  useEffect(() => {
-    const fetchByInputTimer = setTimeout(
-      () => handleFetchMatchingResults(inputString),
-      5000
-    )
-    return () => clearTimeout(fetchByInputTimer)
-  }, [inputString])
-
-  const handleInputChange = useCallback((inputString) => {
-    setInputString(inputString)
-    setIsSubmitted(false)
-  }, [])
-
-  const handleEnterKey = useCallback(
-    (event) => {
-      if (event.key === 'Enter') {
-        setIsSubmitted(true)
-        event.preventDefault()
-      }
-    },
-    [inputString]
-  )
+  const [handleInputChange, handleEnterKey, inputString] = useFetchCharacters()
 
   return (
     <StyledForm>
